Safari-611.1.10/LayoutTests/fast/forms/listbox-respects-padding-bottom-expected.txt

Tests that padding-bottom is respected in non-dropdown listbox'es. Once a listbox is scrolled to its end, clicking on the padding-bottom area should hit an item only after the listbox is scrolled upwards some steps.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS i > 5 && i < maxAttempts is true
PASS successfullyParsed is true

TEST COMPLETE
